+++ lib/pkg_vinculacao.dart
@@ -1,7 +1,6 @@
 library pkg_vinculacao;
 
 export 'src/pages/vinculacao/vinculacao_page.dart';
-export 'src/pages/vinculacao/vinculacao_nova_page.dart';
 export 'src/models/codigo_vinculacao/codigo_vinculacao.dart';
 export 'src/models/dados_aplicativo/dados_aplicativo.dart';
 export 'src/enums/aplicativo_enum.dart';
